// backend/src/models/checkError.ts
export enum CheckError {
  MaxStringLengthError = "max-lenth",
  MinStringLengthError = "min-length",
  ValueNullError = "value-null",
  ValueUndefienedError = "undefiened-null",
  ValueLengthNullError = "value-length-null",
  MinNumberValueError = "min-value",
  MaxNumberValueError = "max-value",
  MinDateValueError = "min-date-value-error",
  MaxDateValueError = "max-date-value-error"
}